+++ src/action-commons/actions/common-sql-rate.adapter.ts
@@ -88,7 +88,9 @@ export class CommonSqlRateAdapter {
             parameters: [].concat(rateUpdateSqlParams).concat(dbId)
         };
 
-        const sqlBuilder = utils.isUndefined(opts.transaction) ? this.knex : opts.transaction;
+        const sqlBuilder = utils.isUndefined(opts.transaction)
+            ? this.knex
+            : opts.transaction;
         const rawUpdate = SqlUtils.executeRawSqlQueryData(sqlBuilder, updateSqlQuery);
         const result = new Promise((resolve, reject) => {
             rawUpdate.then(() => {