+++ src/action-commons/actiontags/common-sql-actiontag-assign.adapter.ts
@@ -123,7 +123,9 @@ export class CommonSqlActionTagAssignAdapter {
             );
         }
 
-        const sqlBuilder = utils.isUndefined(opts.transaction) ? this.knex : opts.transaction;
+        const sqlBuilder = utils.isUndefined(opts.transaction)
+            ? this.knex
+            : opts.transaction;
         const result = new Promise((resolve, reject) => {
             SqlUtils.executeRawSqlQueryData(sqlBuilder, checkBaseSqlQuery).then(dbresults => {
                 const records = this.sqlQueryBuilder.extractDbResult(dbresults, this.knex.client['config']['client']);