+++ src/action-commons/actiontags/common-sql-actiontag-replace.adapter.ts
@@ -149,7 +149,9 @@ export class CommonSqlActionTagReplaceAdapter {
                 ' WHERE ' + replaceConfig.fieldId + '=' + '?' + '',
             parameters: [id]};
 
-        const sqlBuilder = utils.isUndefined(opts.transaction) ? this.knex : opts.transaction;
+        const sqlBuilder = utils.isUndefined(opts.transaction)
+            ? this.knex
+            : opts.transaction;
         const result = new Promise((resolve, reject) => {
             SqlUtils.executeRawSqlQueryData(sqlBuilder, checkBaseSqlQuery).then(dbresults => {
                 const records = this.sqlQueryBuilder.extractDbResult(dbresults, this.knex.client['config']['client']);