+++ docs/CHANGELOG.md
@@ -2,6 +2,22 @@
  
 # Versions
 
+## 5.4.0
+- added support for ordering playlists 
+
+### new features
+- none
+
+### improvements
+- commons: added support for ordering playlists
+
+### bug fixes
+- none
+
+### breaking changes
+- none
+
+
 ## 5.3.0 (2021-05-13)
 - added utils to generate passwords and override options in configfiles